STEP 13. Move out asset actions.
